+++ src/lua_api/remodel.rs
@@ -1,11 +1,14 @@
 use std::{
+    collections::HashMap,
     ffi::OsStr,
     fs::{self, File},
     io::{BufReader, BufWriter},
+    ops::Deref,
     path::Path,
     sync::{Arc, Mutex},
 };
 
+use rbx_dom_weak::{RbxInstanceProperties, RbxTree};
 use rlua::{UserData, UserDataMethods};
 
 use super::LuaInstance;
@@ -14,18 +17,38 @@ pub struct Remodel;
 
 impl UserData for Remodel {
     fn add_methods<'lua, M: UserDataMethods<'lua, Self>>(methods: &mut M) {
-        methods.add_function("readPlaceFile", |_context, lua_path: String| {
+        let master_tree_original = Arc::new(Mutex::new(RbxTree::new(RbxInstanceProperties {
+            name: "REMODEL ROOT".to_owned(),
+            class_name: "REMODEL ROOT".to_owned(),
+            properties: HashMap::new(),
+        })));
+
+        let master_tree = master_tree_original.clone();
+        methods.add_function("readPlaceFile", move |_context, lua_path: String| {
             let path = Path::new(&lua_path);
+
+            let mut master_handle = master_tree.lock().unwrap();
 
             match path.extension().and_then(OsStr::to_str) {
                 Some("rbxlx") => {
                     let file = BufReader::new(File::open(path).map_err(rlua::Error::external)?);
-                    let tree = rbx_xml::from_reader_default(file).map_err(rlua::Error::external)?;
 
-                    let root_id = tree.get_root_id();
-                    let tree = Arc::new(Mutex::new(tree));
+                    let mut source_tree =
+                        rbx_xml::from_reader_default(file).map_err(rlua::Error::external)?;
 
-                    Ok(LuaInstance::new(tree, root_id))
+                    let source_root_id = source_tree.get_root_id();
+                    let source_root = source_tree.get_instance(source_root_id).unwrap();
+                    let source_children = source_root.get_children_ids().to_vec();
+
+                    let master_root_id = master_handle.get_root_id();
+                    let new_root_id =
+                        master_handle.insert_instance(source_root.deref().clone(), master_root_id);
+
+                    for child_id in source_children {
+                        source_tree.move_instance(child_id, &mut master_handle, new_root_id);
+                    }
+
+                    Ok(LuaInstance::new(Arc::clone(&master_tree), new_root_id))
                 }
                 Some("rbxl") => Err(rlua::Error::external(
                     "Reading rbxl place files is not supported yet.",
@@ -37,29 +60,31 @@ impl UserData for Remodel {
             }
         });
 
-        methods.add_function("readModelFile", |_context, lua_path: String| {
+        let master_tree = master_tree_original.clone();
+        methods.add_function("readModelFile", move |_context, lua_path: String| {
             let path = Path::new(&lua_path);
+
+            let mut master_handle = master_tree.lock().unwrap();
 
             match path.extension().and_then(OsStr::to_str) {
                 Some("rbxmx") => {
                     let file = BufReader::new(File::open(path).map_err(rlua::Error::external)?);
-                    let tree = rbx_xml::from_reader_default(file).map_err(rlua::Error::external)?;
+                    let mut source_tree =
+                        rbx_xml::from_reader_default(file).map_err(rlua::Error::external)?;
 
-                    let tree = Arc::new(Mutex::new(tree));
+                    let source_root_id = source_tree.get_root_id();
+                    let source_root = source_tree.get_instance(source_root_id).unwrap();
+                    let source_children = source_root.get_children_ids().to_vec();
 
-                    let instances = {
-                        let tree_handle = tree.lock().unwrap();
+                    let master_root_id = master_handle.get_root_id();
 
-                        let root_id = tree_handle.get_root_id();
-                        let root_instance = tree_handle.get_instance(root_id).unwrap();
-
-                        root_instance
-                            .get_children_ids()
-                            .into_iter()
-                            .copied()
-                            .map(|id| LuaInstance::new(Arc::clone(&tree), id))
-                            .collect::<Vec<_>>()
-                    };
+                    let instances = source_children
+                        .into_iter()
+                        .map(|id| {
+                            source_tree.move_instance(id, &mut master_handle, master_root_id);
+                            LuaInstance::new(Arc::clone(&master_tree), id)
+                        })
+                        .collect::<Vec<_>>();
 
                     Ok(instances)
                 }